Add gRPC service definition to validate an address
This commit introduces a gRPC service called Bitcoin and corresponding
protobuf messages. The definitions can be found in server.proto, and
use the proto3 syntax.

The protobuf messages have been designed to accommodate Bitcoin forks
in future, by using the appropriate network type in ChainParams
message, in a backwards compatible manner.

Finally, we have added a magefile target to generate the protobuf stubs
in Go, like this:
    $ mage proto  # compile .proto files

or, simply do:
    $ mage build  # compile .proto files, then go build