Implement gRPC method to validate addresses
This commit adds the necessary boilerplate to setup a gRPC service for
Bitcoin. As an initial step, the RPC method ValidateAddress has been
implemented. More information can be found in 1c6eec9.

For simplicity, the services do not implement any interface yet, but
should be made more generic to reduce coupling with the main gRPC
server.

The implementation is also accompanied with unit and service level
tests.