Remove lama-protos submodule
It was decided to stop using the lama-protos submodule, in favour of
the previous practice of maintaining .proto files in this repo.

The general thumb rule is to house Protobuf definitions along with the
server-side code, notably for coin services. The .proto files must be
packaged in separate tarballs and published as part of the release.