Decouple gRPC boilerplate from protocol details
This commit is an attempt to ease the process of swapping out the
protocol backend (btcd) with something else.

pkg/bitcoin contains the code that speaks the Bitcoin protocol. It is
the only package that should import btcd/btcutil packages. Protobuf
stubs or gRPC functions should NOT be imported here.

grpc is a new package that contains coin-specific controllers,
implementing the interface required by the gRPC server. This is the
place where we mix together protobuf stubs with protocol-level logic.
Helpers to translate between protobuf-world and pkg world must be added
to the adapters.go file.